Removed copied source files.
All projects now point to a canonical version of the library's code,
which now all resides in MediaParsers.Desktop. This project is what
the MediaParsersTests reads in.

At the same time, removed the duplicate Mp3MediaStreamSource class
from the phone project and aimed its reference at the SL version of
the source code.