Moved the loop which reads the ID3V2 headers to an other thread and made the ReadPastId3v2Tags method asynchronous in order to free the UI thread of the heavy lifting.
An issue with the previous implementation arose when the passed stream was being generated on the fly according to an other NetworkStream, and it seems that even though a Read call is made on a secondary thread, it will still need the UI thread at some point. there appears to be a WaitOne call somewhere and the signal might probably be issued somewhere in the framework on the UI thread.

( see http://nondestructiveme.com/2010/11/23/silverlight-and-networking-the-ui-thread/ )