Fix issue 12: Misses match after partial match
Added bug fix
Added test case

The beginning of an actual match could have been found in a partial match.

For example 001 in 0001. The code attempts to match the 00 but hits the 3rd
zero and drops the partial match. However, it didn't reset itself to the next 0
which is a valid start point for a match and thus 001 is missed in 0001.

See also Issue 12 on github:
https://github.com/loarabia/ManagedMediaHelpers/issues/12